requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Client
httpx==0.25.1

# SOAP Client
zeep==4.2.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Cryptography
cryptography==41.0.7
pyopenssl==23.3.0
python-dateutil==2.8.2

# Caching
cachetools==5.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
